Move context to `tx` package
