+++ integration-tests/testing/chain.go
@@ -22,13 +22,13 @@ import (
 
 // ChainContext is a types used to store the components required for the test chains subcomponents.
 type ChainContext struct {
-	ClientContext config.ClientContext
+	ClientContext tx.ClientContext
 	NetworkConfig config.NetworkConfig
 	keyringMu     *sync.RWMutex
 }
 
 // NewChainContext returns a new instance if the ChainContext.
-func NewChainContext(clientCtx config.ClientContext, networkCfg config.NetworkConfig) ChainContext {
+func NewChainContext(clientCtx tx.ClientContext, networkCfg config.NetworkConfig) ChainContext {
 	return ChainContext{
 		ClientContext: clientCtx,
 		NetworkConfig: networkCfg,
@@ -143,7 +143,7 @@ func NewChain(cfg ChainConfig) Chain {
 	if err != nil {
 		panic(err)
 	}
-	clientContext := config.NewClientContext(app.ModuleBasics).
+	clientContext := tx.NewClientContext(app.ModuleBasics).
 		WithChainID(string(cfg.NetworkConfig.ChainID)).
 		WithClient(rpcClient).
 		WithKeyring(keyring.NewInMemory()).
