+++ pkg/config/network_test.go
@@ -16,6 +16,7 @@ import (
 	"github.com/CoreumFoundation/coreum/app"
 	"github.com/CoreumFoundation/coreum/pkg/config"
 	"github.com/CoreumFoundation/coreum/pkg/staking"
+	"github.com/CoreumFoundation/coreum/pkg/tx"
 	"github.com/CoreumFoundation/coreum/pkg/types"
 	feemodeltypes "github.com/CoreumFoundation/coreum/x/feemodel/types"
 )
@@ -42,7 +43,7 @@ var feeConfig = config.FeeConfig{
 
 func testNetwork() config.Network {
 	pubKey, privKey := types.GenerateSecp256k1Key()
-	clientCtx := config.NewClientContext(app.ModuleBasics)
+	clientCtx := tx.NewClientContext(app.ModuleBasics)
 	tx, err := staking.PrepareTxStakingCreateValidator(clientCtx, ed25519.PublicKey(pubKey), privKey, "1000core")
 	if err != nil {
 		panic(err)
@@ -203,7 +204,7 @@ func TestAddGenTx(t *testing.T) {
 
 	n := testNetwork()
 	pubKey, privKey := types.GenerateSecp256k1Key()
-	clientCtx := config.NewClientContext(app.ModuleBasics)
+	clientCtx := tx.NewClientContext(app.ModuleBasics)
 	tx, err := staking.PrepareTxStakingCreateValidator(clientCtx, ed25519.PublicKey(pubKey), privKey, "1000core")
 	requireT.NoError(err)
 	n.AddGenesisTx(tx)
